Revved rust from 1.62.1 to 1.67.1
- Address Clippy warnings while upgrading from 1.62.1 to 1.67.1
- A fix for the change in signal masking introduced in https://github.com/rust-lang/rust/pull/101077/
- Changes to address an instance of https://rust-lang.github.io/rust-clippy/master/index.html\#result_large_err

+++ components/core/src/os/process/windows_child.rs
@@ -286,7 +286,7 @@ impl Child {
                    stderr: pipes.stderr.map(ChildStderr::from_inner), })
     }
 
-    pub fn id(&self) -> u32 { unsafe { processthreadsapi::GetProcessId(self.handle.raw()) as u32 } }
+    pub fn id(&self) -> u32 { unsafe { processthreadsapi::GetProcessId(self.handle.raw()) } }
 
     pub fn kill(&mut self) -> io::Result<()> {
         cvt(unsafe { processthreadsapi::TerminateProcess(self.handle.raw(), 1) })?;
